Abstracted out Checker and added BranchChecker Moved into a seperate package
