+++ src/main/java/com/lordmat/githubstream/data/checker/CommitChecker.java
@@ -1,6 +1,7 @@
-package com.lordmat.githubstream.data;
+package com.lordmat.githubstream.data.checker;
 
 import com.lordmat.githubstream.bean.GitHubCommit;
+import com.lordmat.githubstream.data.GitHubCaller;
 import com.lordmat.githubstream.util.DateTimeFormat;
 import java.util.Calendar;
 import java.util.Date;
@@ -15,41 +16,24 @@ import java.util.logging.Logger;
  *
  * @author mat
  */
-public class CommitChecker extends Thread {
+public class CommitChecker extends AbstractChecker {
 
     private final static Logger LOGGER = Logger.getLogger(CommitChecker.class.getName());
-
-    /**
-     * How long before checking for new commits
-     */
-    private final static int QUERY_TIME = 30000;
 
     private final GitHubCaller caller;
     private final NavigableMap<Date, GitHubCommit> gitHubCommits;
 
     public CommitChecker(NavigableMap<Date, GitHubCommit> gitHubCommits) {
+        super(30000);
         this.gitHubCommits = gitHubCommits;
         caller = new GitHubCaller();
-    }
-
-    @Override
-    public void run() {
-        while (true) {
-            queryGitHub();
-
-            try {
-                sleep(QUERY_TIME);
-            } catch (InterruptedException ex) {
-                LOGGER.log(Level.FINE, "CommitChecker interrupted", ex);
-            }
-
-        }
     }
 
     /**
      * Handles querying GitHub
      */
-    private void queryGitHub() {
+    @Override
+    protected void query() {
         try {
             String since = null;
